+++ edx_proctoring/serializers.py
@@ -94,7 +94,8 @@ class ProctoredExamStudentAttemptSerializer(serializers.ModelSerializer):
             "id", "created", "modified", "user", "started_at", "completed_at",
             "external_id", "status", "proctored_exam", "allowed_time_limit_mins",
             "attempt_code", "is_sample_attempt", "taking_as_proctored", "last_poll_timestamp",
-            "last_poll_ipaddr", "review_policy_id", "student_name", "is_status_acknowledged"
+            "last_poll_ipaddr", "review_policy_id", "student_name", "is_status_acknowledged",
+            "time_remaining_seconds"
         )