+++ edx_proctoring/tests/test_api.py
@@ -1860,6 +1860,7 @@ class ProctoredExamApiTests(ProctoredExamTestCase):
                 exam_attempt.id,
                 last_poll_timestamp=datetime.utcnow(),
                 last_poll_ipaddr='1.1.1.1',
+                time_remaining_seconds=600,
                 status='foo'
             )
 
@@ -1868,12 +1869,14 @@ class ProctoredExamApiTests(ProctoredExamTestCase):
             exam_attempt.id,
             last_poll_timestamp=now,
             last_poll_ipaddr='1.1.1.1',
+            time_remaining_seconds=600,
         )
 
         attempt = get_exam_attempt_by_id(exam_attempt.id)
 
         self.assertEqual(attempt['last_poll_timestamp'], now)
         self.assertEqual(attempt['last_poll_ipaddr'], '1.1.1.1')
+        self.assertEqual(attempt['time_remaining_seconds'], 600)
 
     def test_requirement_status_order(self):
         """